Level, Tiles, Sprites.
fixed our image IO in SpriteSheet and added that sample sheet as a
spritesheet
Added a grass sprite and tile (sprite not working yet, shows different
part of spritesheet).
Added a sprite variable to Tile, a set of static final ints that
identify tiles like enums, and a set of static final ints that identify
tiles based on ARGB values from pixels
Added level loading and generation to Level
Added renderSprite to screen

Overall, made it so the level actually draws by fixing ImageIO and
adding a sprite, level generation, and tile rendering.