Sprites work! And other good changes
Biggest change is getting sprites to work. They work now thanks to the
fact that it now multiplies our "y" in load() in Sprite by the sprite
sheet's width.
Other changes:
Got rid of the arraycopy thing going on in Sprite. Replaced it with an
arraylist instead of removing it in case there were ambitions for it I
don't know about.
Player renders in the middle of the screen now
changed renderPlayer method to call renderSprite because they're the
exact same
I replaced size with width and height everywhere.
I made it so levels aren't just grassTiles but also voidTiles in
generateLevel()
I added an isometric office tiles image to our tile textures pack.
Hopefully we can make this game isometric. It's only a minor rendering
difference: render from the top down. left to right sprites with
transparent corners.