Isometric and more
Biggest change is making rendering isometric
Also created a flipped sprite method that is used to create another
sprite for the player when they go right or up
Created isoTo2d to twoDtoIso to convert between Xs and Ys in each of
them.
Created a RaisedObject class for walls, desks, etc. that should be
rendered above tiles and integrated it into level and screen
Loaded the sprites from the office sprite pack
Still need to: get RaisedObject rendering to work, fix the cubicles
spritesheet since all of its sprites are off origin