Small fixes
Added a way to randomly change the text being sent to the display text
class, just press 'p' on the keyboard. Fixed a bit of the commenting in
the main class and screen. Made code look a bit nicer through removal of
commented out lines and out-of-place variables in Screen. Added a little
piece of text that says "Press x or space..." whenever the textbox
appears. A little more work on getting additional screens to work.