Blur
Added a blur that could be used in full screen because of jagged lines,
but left it commented out. It works but its not anti-aliasing and its
slow. More work needed if it's going to be a thing. Possible algorithms
to use for anti-aliasing are the ones retro game emulators use like hqx
or brz.