HUD
BusinessSim.render now calls HUD.displayHUD(Graphics g). The HUD
displays the FPS, current floor, and items left for pickup. Noticed
issue: in fullscreen the HUD goes from the left of the screen to the
right of the game screen. I bet it's the same issue we have with our
lines that connect the player to pickups.