Minor adjustments
HUD says when you're near a pickup, text displayer is scaled (and
broken? it's not showing second page of text despite me not changing
anything relevant to that in this update. Maybe I just haven't noticed
it). Tried doing a bunch of stuff to text displayer but it wasn't
working, will have to fix it next update possibly with a revamp of
textDisplayer.