Added level5/6 files, npc sayings files
I created level5 and 6 object and floor files, coped from floors 1 and
2.
I created NPC sayings files.
I moved questions and npc sayings into a text folder.
I created a profiler for NPCs.
I edited readFiles in NPC to look nicer although it became slower
according to the profiler, meaning we might want to reverse some of the
changes to that method that this push includes.