Working elevator
Elevator has been revamped to support mouse input. Now it's just up and
down buttons.
I changed the rendering of the panel and the logic of the panel. It
registers the elevator as pointing when you're hovering over the option
but only does something when mouse.lastMouseClicked is true.
I made sure to make the new rendering include various color clues to the
user: the buttons change color when you hover over them and when you
hold them.