Update Questions
I changed the question system to accept multiple questions per topic,
though there has to be the same number of questions per topic. I also
changed how the HuntObject receives the question. The reasoning is now
tacked onto the end of the question text, but is separated by the
HuntObject when it receives the information.
Floor one is updated, but 2-5 are not. The game will not run from this
update, but I feel it is critical point to push. I also moved Dialogue
to util because I feel it belongs there more since it doesn't handle
anything Graphical. Ciao